Clarify config of bingo using $GOBIN
This commit changes the recommended way to set up the environment for bingo to
work as expected after:

https://github.com/bwplotka/bingo/blob/bc83849427217f3408b91d98edb0a3b8fce18e9e/README.md#using-installed-tools

And in the source code:

	// gobin mimics the way go install finds where to install go tool.
	func gobin() string {
		binPath := os.Getenv("GOBIN")
		if gpath := os.Getenv("GOPATH"); gpath != "" && binPath == "" {
			binPath = filepath.Join(gpath, "bin")
		}
		return binPath
	}

Asking users to change the GOPATH was misleading because it is a variable that
was mostly used in pre-go1.13 times when we didn't have Go modules. Also,
`go env GOPATH` defaults to `$HOME/go` when undefined.