Make prefix an attribute of the Redis store
The ticket provider used to generate prefixed keys. This commit pushes that
concern out to the ticket store since namespacing is a property managed
differently depending on the backend, e.g. in-mem does not need it, other
store may not namespace in a string-based form, e.g. multi-dimensional arrays,
tables, etc...