Set CORS Allow-Origin header by config
Allow configuration of the CORS "Access-Control-Allow-Origin" header in
the API from a config file or an environment variable.

If `CORSOrigin` is not set in the config file or an Envvar, then it will
be set to the domain of the Enduro API (e.g. "http://localhost:9000"),
effectively disabling cross-origin access. This default behaviour
prevents Enduro from exiting when `CORSOrigin` is not set, while
preventing unexpected CORS access to the API endpoints.