Add bucket package
This commit introduces a new package `bucket` aiming to consolidate the
different ways we configure and create buckets across the application.

It enables future work such as embedding the shared config in different
contexts as well as using memory-based or fs-based buckets for testing purposes
when setting up some components of the application.

Relates to https://github.com/artefactual-sdps/enduro/issues/858.