Add a rand parameter to the storage service
This modification enables the use of a custom random number generator in our
tests. It eliminates the need to alter global state in the
`github.com/google/uuid` package, which is unsafe during parallel testing.

We continue to utilize `crypto/rand`, but now it must be explicitly passed to
the service. We maintain the assumption that the operation will succeed, and
will trigger a panic otherwise, to mirror the behavior of `uuid.New`. We have
implemented middleware to manage panics, thus mitigating concerns in the
unlikely event of such failures.