Use OneOf in monitor events
OneOf is the DSL in Goa to express union types, which we use in this commit to
describe the different type of event attributes that we expect in MonitorEvent.

The JSON encoding of union types is different enough that significant changes
were produced by the OpenAPI-generated client and our client code in the
dashboard. It also impacted the Redis-based event broker which relied on the
JSON-encoded form of the events to stream the data over the wire.