Fix AM transfer deletion
The SFTP client uses a `remoteDir` config value that is used as prefix
in the `Upload()` and `Delete()` functions. Using the remote path from
the upload transfer result duplicates that prefix and makes the deletion
to fail with a "file does not exist" error.

Return also the remote relative path from the upload activity and use it
as destination for the delete activity.

Refs #785.