Fix "can't find bundle activity" error
Refs #795.

I haven't experienced the periodic "can't find bundle activity" error
in testing this change, but it's hard to know if it's a fix due to the
unpredictable nature of the bug. Even if these change doesn't fix the
bug, the additional logging and tests should be helpful for diagnosing
the problem.

- Use an intermediate variable to assign the bundle activity results to
  tinfo.Bundle
- Add logging to the bundle activity and log the execution params
- Update workflow tests
- Add bundle activity unit tests (39% -> 64% coverage in workflow
  activities 🎉)
- Generate a mock watcher for tests