Use location and PackageReader for Download
This implementation does not use the preservation config for determining
the correct call to download, instead it leverages the package reader
function and the Location interface to manage the correct storage service
to use.

This is partially implemented. Not sure how to actually snag the url from
the location configuration. I really need guidance there. This also seems
rather complex compared to the other implementation.