Enable tracing in http.Client
Register a propagator with OpenTelemetry to enable propagation of trace context
over the wire. Update HTTP clients with `otelhttp.Transport` to wrap outbound
requests with a span and enrich it with metrics.

This change enable us to relate spans occuring beyond an individual service,
e.g. the activity "move-to-permanent-storage" executed by enduro-a3m-worker uses
a HTTP client to request to the Storage Service the long-term storage of the
AIP. With this commit we'll start seeing the `HTTP POST` operation performed by
the client but also the rest of the details happening on the server side (child
spans), e.g.: API handler, SQL queries, workflow, activities, functions...