Fix user hydration when authenticated
After updating to Vue 3.4, the global navigation guard in the router stopped
being able to validate authenticated users during the initial loading of the
application, e.g. after a page refresh. The root issue lies in the `getUser`
success callback in `main.ts` executing too late, i.e. not having run by the
time when the navigation guard was evaluating `isUserValid`. I'm not sure what
is the exact change in Vue that caused this issue, but I suspect it's related
to the updates in the reactivity system.

The fix proposed consists in loading the user from the layout store using a new
action that is triggered by the navigation guard.

The `App` component now uses `watch` to observe `layoutStore.isUserValid` and
react by establishing the package monitor connection when it's set to true,
effectively decoupling the connection process from the authentication workflow.