Add a persistence query filter implementation
Refs #988, #514

To address #988 and provide a count of the total number of packages in
the dashboard the API GET /package endpoint to provide total number of
packages in a result set. Because we need to update the API and the
underlying database calls, this is a good opportunity to also address
issue #514 and add a general search implementation that allows
filtering, ordering and paging results. This PR adds this general search
filtering implementation with paging and ordering, and also provides a
count of the total number of records matching the search criteria before
paging.

- Add a persistence package filter stub implementation
- Add persistence Page and Sort structs
- Add an entgo query filter implementation
- Add methods to add page, sort, and column value filters to the ent
  filter implementation
- Add a timerange package for representing time ranges (i.e. a time
  period with a start and end time)