Copy premis.xsd into enduro image
This file should only be needed by the AM and a3m workers to do PREMIS
validation when it's enabled. However, config validation is the same for
all commands and, when PREMIS validation is enabled, it checks the
existence of the XSD file in the filesystem. We tried to disable PREMIS
validation only in the enduro container, but that config value is read
in the workflow context, disabling the scheduling of the activity from
the workflow worker.