Add created time package list filter, refs #1102
Add a package list dropdown filter to find packages created recently
(in the last 3 hours to 7 days). I still need to do work to collapse
the status tabs into a dropdown list, add the hidden filter panel,
etc. to fully implement #1102, but I think this a good point to
review the code and functionality.