Add AIP deletion workflow
Add a two steps process to trigger the deletion of an AIP. Create AIP
deletion-request and deletion-review endpoints in the storage service
API. The request will start a Temporal workflow using a deterministic
workflow ID based on the AIP UUID, this ID can only be reused when the
workflow finishes in a failed status (error, canceled, terminated) but
not when a previous workflow succeeds or is currently running. The
workflow will stop and wait for a review signal that is sent from the
API review workflow. The request requires a reason and the review can
not be performed by the requester. Both endpoints require their own
ABAC scopes.

The deletion workflow will look at the AIP and related location info and
it will work with MinIO and AMSS source locations. For the AMSS deletion
there are a few requests involved, to get the origin pipeline from the
AIP details, which is needed to request the deletion, and then it will
poll for status until it changes from "DEL_REQ" to "UPLOADED" or
"DELETED". Approval in the AMSS must be done manually or with a custom
script, since there is no way to approve the deletion in the API yet.

Add DeletionRequest schema and DB table to hold information about the
request/review and new AIP and Workflow statuses. Add persistence
methods to create and update deletion requests and to get the currently
pending request.

Add delete button, request dialog with reason textarea and review alerts
to the dashboard. Allow task card note updates and render emails in them
as mailto links. Add a poll process after request and review to update
the current status until we implement websockets in the storage service.
Change status filters and time dropdown label in the AIPs browse page
and switch status badge background colors between "queued"
and "processing"/"in progress" statuses. Improve buttons behavior in
the location card and remove move button for now.

[skip-codecov]