Fix usage of golangci-lint
Previously we relied on `golangci-lint fmt` in CI, but it only reformats code
without ever returning a non‐zero exit code on violations. By switching to
`golangci-lint run`, we get both formatter and linter checks in one command

We use `--fix` locally (via make lint) to automatically correct issues during
development, but disable it in CI (`--fix=0`) so that the pipeline reports
problems without mutating the code.

We can still use `make fmt` (equals to `golangci-lint fmt`) if you need
faster formatting, as it does not invoke the linters.