Add download SIP API endpoints
Browsers don't allow to set custom headers (like Authorization) when
downloading files via `window.open()` or `<a href="...">`. Also, using
the fetch API with custom headers requires loading the entire file into
memory, which is not suitable for large files. To work around this, I
implemented a two-step download process, similar to what we do to
set up Websocket connections:

- Request SIP download endpoint:
  - The frontend will call this endpoint with the Authorization header.
  - The backend will validate the token and return a short-lived,
    single-use ticket (via a cookie).

- SIP download endpoint:
  - The frontend will open this endpoint in a new window.
  - The browser will automatically send the ticket cookie.
  - The backend will validate the ticket and stream the file.

Following iterations could allow to download the SIP in a single step
process for out of browser requests to the API.