Allow saving key/value in the auth ticket store
Gives the option to store a value associated with a generated ticket.
The ticket/key is sent as a cookie while the value is saved in the store
and scanned when the cookie is received. This will allow to save the
user claims between the monitor request and the monitor endpoints.