Fix API schema generation
Goa considered StorageMonitorEvent and IngestMonitorEvent
structurally equivalent which prevented the inclusion of the latest in
the schema. Adding the "openapi:typename" meta makes sure both are
generated but then Goa creates HTTP and Websocket schema versions,
calling the latest `StorageMonitorEvent2`. By changing the `event` field
to use different names on each type we get a single event type in the
schema for each service.