Allow SIP source upload from the dashboard
Split the upload SIPs page into local and cloud upload using tabs. Move
the existing Uppy uploader to a component and create a new one for the
SIP source upload. In the new component, fetch a list of objects from a
hardcoded SIP source using scroll load and display a list-group with
check-boxes and multi-selection. Checking the objects and submitting
the upload will group and make all requests to the API to start an ingest
workflow for each SIP.

Co-authored-by: David Juhasz <djjuhasz@gmail.com>