Fix uploader creation on SIP upload
Use a single transaction to find or create the user and create the SIP.
Reuse the User datatype for the SIP Uploader to be able to pass the
OIDC fields on SIP creation. Also, check user claims before streaming
the upload and rename SIP edge in the DB schema from to uploader.