Start ingest workflow from SIP source
Add API endpoint to create a SIP from a SIP source and init the ingest
workflow. Add `SIPSourceID` field to the workflow request and register
two `bucketdownload` activities, one using the internal bucket and the
other using the SIP source bucket. Based on the `SIPSourceID` field run
one activity or the other.

Co-authored-by: David Juhasz <djjuhasz@gmail.com>