Fix user creation on concurrent add SIP requests
Add unique index over the user OIDC columns combined to prevent
the creation of multiple users with the same OIDC iss and sub claims.
To allow setting that index without prefix I reduced both column sizes
to 255 chars.

Additionally, when multiple requests are sent at the same time by a new
user, most of them would fail due to the new unique index. To allow that
scenario I decided to use Ent's upsert feature. This will create a new
user or update the email and name of an existing one if there is a
conflict in the OIDC columns.