Add config docs to admin manual (#1336)
Adds a new page to the Administrator's manual covering configuration. Currently just focused on the main enduro.toml file but can be expanded for other config files if that becomes relevant in the future. Also adds or adjusts some of the comments on the TOML file itself. Squashes in a lot of good PR feedback fixes as well.