Add a "Copy SIP" workflow task
Refs #1289.

Add a "Copy SIP" task to the "Create AIP" workflow that provides
early feedback to the operator about the progress of the workflow.
The "Copy SIP" task will be displayed on the SIP detail page right
after the SIP is uploaded, which will also make the "Create AIP"
workflow show up sooner.