Reconnect websockets when they close unexpectedly
Fixes #1057.

Attempt to reconnect the ingest and storage monitor websockets when the
connection is unexpected closed at the server end (e.g. on server
restart). My tests so far indicate that this change fixes the problem
I've been having with the Dashboard hanging intermittently and throwing
Vite "write EPIPE" errors.