Add retention period to all SIP origin configs
Allow setting a retention period to the internal uploads and the SIP
sources, and update the current behavior in watched buckets for
consistency. The retention period will default to "0" (automatic
deletion) and has to be set to a negative value to disable the deletion
of the original SIP after a successful ingest. Record a workflow task to
show details in the API/dashboard.

[skip-codecov]