Fix WebSocket events parsing in dashboard
Goa now generates two OpenAPI schemas for the storage and ingest event
types even if Meta("type:generate:force") is not set. We need to use the
IngestEvent2 and StorageEvent2 types in the dashboard to parse. Also,
remove metas from events types; members of OneOf are now included in the
schema, even if they are structurally equivalent. We can now use the
same value property name for both event types.

[skip-codecov]