Fix AIP Deletion Report data races
Refs #1207.

Fix two data races when calling the pdfcpu `FillForm()` from
multiple threads.  The first race occurred when multiple threads tried
to read or write to the pdfcpu configuration directory, which is not
easily configurable.  The second data race occurred when multiple
threads tried to mutate the shared `model.Configuration` struct.

Note that the `pdf.NewPDFCPU()` constructor is still not thread-safe,
due to the need for a shared configuration directory, but we don't need
to call it from multiple threads with this code.

- Create a `pdf` package to better isolate the pdf editing code
- Create a `pdf.FormFiller` interface to abstract the pdfcpu
  implementation and allow mocking the implementation in test
- Inject a `pdf.FormFiller` implementation in the
  `NewAIPDeletionReport()` constructor
- Move all the report generation code to the `AIPDeletionReportActivity`
  to reduce complexity and data passing
- Remove the `reports` package