Increase default AM and a3m capacity
To allow processing SIPs from batches in parallel we need to increase
the a3m and AM workers capacity. Use the default limits from Temporal
for:

- MaxConcurrentSessionExecutionSize (a3m and AM)
- MaxConcurrentActivityExecutionSize (AM)

Use the capacity configuration to set both values to allow concurrent
sessions executing activities in parallel in the AM worker. Keep the
a3m worker limited to a single activity as a3m fails with concurrent
transfers.