Add SBOM GitHub Action workflow
Add new GitHub Action workflow that will generate an SBOM file using Trivy
and upload it to Dependency Track.

This workflow will run on push to main, and when a new release is created.

This workflow requires the secrets DEPENDENCY_TRACK_URL and
DEPENDENCY_TRACK_API_KEY be configured in GitHub to allow upload to
Dependency Track.