Release session while waiting for batch signal
Complete the current session before waiting for the batch signal to
allow other sessions to execute in the worker. Recreate the session
after in the same worker to continue having access to the local files.

The recreated session takes an activity slot in the same worker where
the functional activities run. Set `MaxConcurrentActivityExecutionSize`
to the double of `MaxConcurrentSessionExecutionSize` in both workers
to avoid deadlocks in this situation.

[skip-codecov]